8377658: G1: Cleanup g1ConcurrentMark.cpp for stricter warning flags
Cleanup g1ConcurrentMark.cpp and included header files to remove
implicit narrowing conversions.

For gcc: `-Wconversion -Wno-float-conversion`
For clang: `-Wimplicit-int-conversion`

+++ src/hotspot/share/gc/g1/g1RegionMarkStatsCache.hpp
@@ -98,7 +98,7 @@ private:
   size_t _num_cache_entries_mask;
 
   uint hash(uint idx) {
-    return idx & _num_cache_entries_mask;
+    return idx & static_cast<uint>(_num_cache_entries_mask);
   }
 
   G1RegionMarkStatsCacheEntry* find_for_add(uint region_idx);
